Move query param helpers into editor package
