+++ packages/example-app-base/src/index2.ts
@@ -22,4 +22,3 @@ export * from './components/Table/index';
 export * from './components/TopNav/index';
 export * from './components/Video/index';
 export * from './utilities/index2';
-export * from './utilities/settings';
