+++ packages/example-app-base/src/utilities/index2.ts
@@ -5,7 +5,6 @@ export * from './appInsightsHelper';
 export * from './baseDefinition';
 export * from './getEditUrl';
 export * from './getPageFirstPlatform';
-export * from './getQueryParam';
 export * from './getSiteArea';
 export * from './jumpToAnchor';
 export * from './PlatformContext';
@@ -15,3 +14,4 @@ export * from './removeAnchorLink';
 export * from './SiteDefinition.types';
 export * from './string';
 export * from './windowWidth';
+export { getQueryParam, getSetting } from '@uifabric/tsx-editor/lib/utilities/index';
