+++ packages/tsx-editor/src/index.ts
@@ -1,3 +1,5 @@
-import './version';
 export * from './components/index';
 export * from './transpiler/index';
+
+// Intentionally not exporting the utilities, since those really ought to live in example-app-base
+// and are exported from there
